move FederationStream out to its own file
